Complete day 10 part 1
More luck than anything else. Ended up implementing a size check to only output the message once it got to a size range. Got lucky on the first try with 30 but spent a while before this trying to figure out how to reduce size in a reasonable way. As the starting data is massive and can not be outputted