Completed Day 11 part 1
Very easy, most of the work was just typing in the rules. Over all it didn't seem hard but I think this is due to my experience with this type of problem. A while back I did something like this for Minecraft crafting grids. In which I need to take a 7x7 and match it to any 3x3 recipe. Same general concept minus the long list of math to get the power level.