Completed day 12 part 1
Even though I completed this part my solution is not idea. As it requires a user to input the number of pots to allow plants to grow into. Looking at part 2 this will never work as it wants a 50B generation solution. Which could result in infinite growth in both directions. This means my next solution will need a linked list of something of that nature.