Completed day 12 part 2
This is not an actual solution as it will not give an answer. Instead how it work is to allow running larger sets of data then the first version. Which turned out to be useless as I don't have enough CPU nor RAM to run this to 50B. Instead I found a reddit post that suggested looking for patterns. To get a solution you find the pattern of repeat in the deltas. Then use the equation '(50B - repeatIndex) * repeatDelta + repeatSum' to get the answer. In my case the answer was 4,300,000,002,414